requirements.txt
pandas~=2.1.1
numpy~=1.23.1
scikit-learn~=1.1.1
joblib~=1.2.0
flask~=2.2.2
pymongo~=4.2.0
python-dotenv~=0.21.0
requests~=2.31.0